requirements.txt
streamlit>=1.36.0
requests>=2.32.3
python-dotenv>=1.0.1
ollama>=0.1.8
pytest>=8.2.0
